Migrate chapters 3 & 4
