+++ src/applications/hca/config/chapters/householdInformation/dependentInformation.js
@@ -1,0 +1,41 @@
+import { assign } from 'lodash/fp';
+import fullSchemaHca from 'vets-json-schema/dist/10-10EZ-schema.json';
+
+import DependentView from '../../../components/DependentView';
+import { uiSchema as dependentUI } from '../../../definitions/dependent';
+
+const { dependents } = fullSchemaHca.properties;
+
+export default {
+  path: 'household-information/dependent-information',
+  title: 'Dependent information',
+  depends: data => data.discloseFinancialInformation,
+  uiSchema: {
+    'view:reportDependents': {
+      'ui:title': 'Do you have any dependents to report?',
+      'ui:widget': 'yesNo',
+    },
+    dependents: {
+      items: dependentUI,
+      'ui:options': {
+        expandUnder: 'view:reportDependents',
+        itemName: 'Dependent',
+        hideTitle: true,
+        viewField: DependentView,
+      },
+      'ui:errorMessages': {
+        minItems: 'You must add at least one dependent.',
+      },
+    },
+  },
+  schema: {
+    type: 'object',
+    required: ['view:reportDependents'],
+    properties: {
+      'view:reportDependents': { type: 'boolean' },
+      dependents: assign(dependents, {
+        minItems: 1,
+      }),
+    },
+  },
+};
